backend/trading_algorithm/requirements.txt
web3==6.11.1
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
matplotlib==3.8.2
tatumio==1.3.1
aiohttp==3.9.1
websockets==12.0
backoff==2.2.1
prometheus-client==0.19.0
structlog==24.1.0
tenacity==8.2.3
aiodns==3.1.1
cchardet==2.1.7
uvloop==0.19.0; sys_platform != 'win32'  # Performance boost for asyncio
numpy==1.26.3
scipy==1.12.0
statsmodels==0.14.1  # For advanced statistical analysis
aioredis==2.0.1  # For caching
orjson==3.9.10  # Faster JSON processing
prometheus-async==22.2.0  # Async support for Prometheus
python-json-logger==2.0.7  # JSON logging
pytz==2024.1  # Timezone support 